+++ tests/test_multilabel.py
@@ -18,7 +18,6 @@ GlobalAveragePooling1D = ktrain.imports.keras.layers.GlobalAveragePooling1D
 
 
 def synthetic_multilabel():
-
     # data
     X = [
         [1, 0, 0, 0, 0, 0, 0],
@@ -93,7 +92,7 @@ class TestMultilabel(TestCase):
         # use loss instead of accuracy due to: https://github.com/tensorflow/tensorflow/issues/41114
         hist = learner.fit(0.001, 200)
         learner.view_top_losses(n=5)
-        learner.validate()
+        print(learner.validate())
         # final_acc = hist.history[VAL_ACC_NAME][-1]
         # print('final_accuracy:%s' % (final_acc))
         # self.assertGreater(final_acc, 0.97)